TF-1427 Separate folder for email search
(cherry picked from commit 373405e4433a2532c4df42b51c09f715bc8db695)
