+++ lib/features/search/email/presentation/model/search_more_state.dart
