+++ lib/features/search/email/presentation/search_email_controller.dart
@@ -15,7 +15,16 @@ import 'package:jmap_dart_client/jmap/core/utc_date.dart';
 import 'package:jmap_dart_client/jmap/mail/email/email_address.dart';
 import 'package:jmap_dart_client/jmap/mail/email/email_comparator.dart';
 import 'package:jmap_dart_client/jmap/mail/email/email_comparator_property.dart';
-import 'package:model/model.dart';
+import 'package:model/email/email_action_type.dart';
+import 'package:model/email/mark_star_action.dart';
+import 'package:model/email/prefix_email_address.dart';
+import 'package:model/email/presentation_email.dart';
+import 'package:model/email/read_actions.dart';
+import 'package:model/extensions/list_presentation_email_extension.dart';
+import 'package:model/extensions/presentation_email_extension.dart';
+import 'package:model/mailbox/presentation_mailbox.dart';
+import 'package:model/mailbox/select_mode.dart';
+import 'package:model/user/user_profile.dart';
 import 'package:tmail_ui_user/features/base/base_controller.dart';
 import 'package:tmail_ui_user/features/contact/presentation/model/contact_arguments.dart';
 import 'package:tmail_ui_user/features/destination_picker/presentation/model/destination_picker_arguments.dart';
@@ -35,11 +44,11 @@ import 'package:tmail_ui_user/features/mailbox_dashboard/domain/usecases/save_re
 import 'package:tmail_ui_user/features/mailbox_dashboard/presentation/model/dashboard_routes.dart';
 import 'package:tmail_ui_user/features/mailbox_dashboard/presentation/model/search/email_receive_time_type.dart';
 import 'package:tmail_ui_user/features/mailbox_dashboard/presentation/model/search/quick_search_filter.dart';
-import 'package:tmail_ui_user/features/search/domain/state/refresh_changes_search_email_state.dart';
-import 'package:tmail_ui_user/features/search/domain/usecases/refresh_changes_search_email_interactor.dart';
-import 'package:tmail_ui_user/features/search/presentation/model/search_more_state.dart';
-import 'package:tmail_ui_user/features/search/presentation/model/simple_search_filter.dart';
-import 'package:tmail_ui_user/features/search/presentation/search_email_bindings.dart';
+import 'package:tmail_ui_user/features/search/email/domain/state/refresh_changes_search_email_state.dart';
+import 'package:tmail_ui_user/features/search/email/domain/usecases/refresh_changes_search_email_interactor.dart';
+import 'package:tmail_ui_user/features/search/email/presentation/model/search_more_state.dart';
+import 'package:tmail_ui_user/features/search/email/presentation/model/simple_search_filter.dart';
+import 'package:tmail_ui_user/features/search/email/presentation/search_email_bindings.dart';
 import 'package:tmail_ui_user/features/thread/domain/constants/thread_constants.dart';
 import 'package:tmail_ui_user/features/thread/domain/model/search_query.dart';
 import 'package:tmail_ui_user/features/thread/domain/state/empty_trash_folder_state.dart';
