+++ lib/features/search/email/presentation/search_email_view.dart
@@ -18,13 +18,13 @@ import 'package:tmail_ui_user/features/mailbox_dashboard/presentation/model/sear
 import 'package:tmail_ui_user/features/mailbox_dashboard/presentation/model/search/quick_search_filter.dart';
 import 'package:tmail_ui_user/features/mailbox_dashboard/presentation/widgets/email_quick_search_item_tile_widget.dart';
 import 'package:tmail_ui_user/features/mailbox_dashboard/presentation/widgets/recent_search_item_tile_widget.dart';
-import 'package:tmail_ui_user/features/search/presentation/model/search_more_state.dart';
-import 'package:tmail_ui_user/features/search/presentation/model/simple_search_filter.dart';
-import 'package:tmail_ui_user/features/search/presentation/search_email_controller.dart';
-import 'package:tmail_ui_user/features/search/presentation/utils/search_email_utils.dart';
-import 'package:tmail_ui_user/features/search/presentation/widgets/app_bar_selection_mode.dart';
-import 'package:tmail_ui_user/features/search/presentation/widgets/email_receive_time_action_tile_widget.dart';
-import 'package:tmail_ui_user/features/search/presentation/widgets/email_receive_time_cupertino_action_sheet_action_builder.dart';
+import 'package:tmail_ui_user/features/search/email/presentation/model/search_more_state.dart';
+import 'package:tmail_ui_user/features/search/email/presentation/model/simple_search_filter.dart';
+import 'package:tmail_ui_user/features/search/email/presentation/search_email_controller.dart';
+import 'package:tmail_ui_user/features/search/email/presentation/utils/search_email_utils.dart';
+import 'package:tmail_ui_user/features/search/email/presentation/widgets/app_bar_selection_mode.dart';
+import 'package:tmail_ui_user/features/search/email/presentation/widgets/email_receive_time_action_tile_widget.dart';
+import 'package:tmail_ui_user/features/search/email/presentation/widgets/email_receive_time_cupertino_action_sheet_action_builder.dart';
 import 'package:tmail_ui_user/features/thread/domain/state/search_email_state.dart';
 import 'package:tmail_ui_user/features/thread/domain/state/search_more_email_state.dart';
 import 'package:tmail_ui_user/main/localizations/app_localizations.dart';
