+++ lib/features/search/email/presentation/utils/search_email_utils.dart
